.git-rewrite/t/requirements_enhanced.txt
# ULTRON Agent 3.0 - Enhanced Requirements
# Core dependencies with security and performance improvements

# Core AI and Voice
pyttsx3>=2.90
speechrecognition>=3.10.0
openai>=1.0.0
whisper>=1.0.0

# Security and Encryption
cryptography>=41.0.0
pyjwt>=2.8.0

# Web and API
requests>=2.31.0
flask>=3.0.0
flask-socketio>=5.3.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0
websockets>=12.0
httpx>=0.27.0

# Database and Storage
supabase>=2.0.0
faiss-cpu>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=4.6.0

# System and Performance
psutil>=5.9.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
celery>=5.3.0

# GUI and Multimedia
pygame>=2.5.0
pillow>=10.0.0
pynput>=1.7.6
pyautogui>=0.9.54
opencv-python>=4.8.0
PySide6>=6.7.0
pyaudio>=0.2.14

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Machine Learning (Optional)
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# Blockchain and Web3 (Optional)
web3>=6.10.0
solana>=0.30.0
qiskit>=0.44.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# Logging and Monitoring
loguru>=0.7.0
rich>=13.5.0

# CLI and Utilities
typer>=0.9.0
streamlit>=1.26.0
pydantic>=2.3.0

# Audio Processing
elevenlabs>=0.2.0
python-socketio>=5.11.2

# Additional utilities
Jinja2>=3.1.2
gputil>=1.4.0
pytesseract>=0.3.10
asyncio-mqtt>=0.16.0

# Security scanning (development)
bandit>=1.7.5
safety>=2.3.0